requirements.txt
pymorphy2
asgiref==3.3.1
bert-extractive-summarizer==0.5.1
blis==0.7.4
boto3==1.16.32
botocore==1.19.32
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cymem==2.0.5
Django==3.1.4
filelock==3.0.12
fr-core-news-md==2.3.0
idna==2.10
it-core-news-md==2.3.0
jmespath==0.10.0
joblib==0.17.0
lxml==4.6.2
murmurhash==1.0.5
neuralcoref==4.0
nltk==3.5
numpy==1.19.4
packaging==20.7
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytils==0.3
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.94
six==1.15.0
spacy==2.3.4
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.4
threadpoolctl==2.1.0
tokenizers==0.9.4
tools==0.1.9
torch==1.4.0
torchvision==0.5.0
tqdm==4.54.1
transformers==2.2.2
urllib3==1.26.2
wasabi==0.8.0
